pyproject.toml
[project]
name = "mini-model-serving-platform"
version = "0.1.0"
description = "Mini internal ML serving platform with registry, Triton serving, canary deploy, monitoring, and GitOps."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "MMSP", email = "ml-platform@example.com" }]
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.27.1",
  "typer==0.9.0",
  "pydantic==2.6.3",
  "pydantic-settings==2.2.1",
  "requests==2.31.0",
  "httpx==0.27.0",
  "prometheus-client==0.20.0",
  "numpy==1.26.4",
  "pandas==2.2.1",
  "scipy==1.12.0",
  "pyyaml==6.0.1",
  "python-json-logger==2.0.7",
  "tritonclient[http]==2.41.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.1.1",
  "pytest-asyncio==0.23.5",
  "ruff==0.3.5",
  "mypy==1.8.0",
  "types-PyYAML==6.0.12.12",
]

[project.scripts]
mmsp = "mmsp.cli:app"

[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true
